add dashboard API endpoints and disable registration
- Add dashboard API methods for stats, recent payments, overdue members, and recent activities
- Disable user registration in backend with 403 response and remove frontend registration routes
- Add validation annotations to Member and Payment domain models
- Change Payment amount from BigDecimal to Double for simplicity
- Update development server port to 8082 and adjust frontend proxy configuration
- Remove registration link from login view and update messaging

+++ src/main/java/io/github/membertracker/domain/exception/PaymentDomainException.java
@@ -1,6 +1,5 @@
 package io.github.membertracker.domain.exception;
 
-import java.math.BigDecimal;
 import java.time.YearMonth;
 
 /**
@@ -27,7 +26,7 @@ public class PaymentDomainException extends DomainException {
     }
 
     // Factory methods for common payment domain violations
-    public static PaymentDomainException invalidPaymentAmount(BigDecimal amount, BigDecimal minAmount) {
+    public static PaymentDomainException invalidPaymentAmount(Double amount, Double minAmount) {
         return new PaymentDomainException(
             String.format("Payment amount %s is invalid. Minimum amount is %s", amount, minAmount),
             INVALID_PAYMENT_AMOUNT
